Refactor so full PHP version is generated once and used everywhere else

+++ php/scripts/build.sh
@@ -37,11 +37,12 @@ IMAGE_NAME=${IMAGE_NAME:-php}
 if [[ "${INDEV_FLAG:-1}" != "0" ]]; then
   IMAGE_NAME="${IMAGE_NAME}-indev"
 fi
-  MAJOR_VERSION="$(echo "${BUILD_VERSION}" | sed -E 's/([0-9])([0-9])/\1.\2/')"
-  # Configure build args specific to this image build
-  export PHP_VERSION="${MAJOR_VERSION}"
-  BUILD_ARGS=(PHP_VERSION)
-  BUILD_ARGS+=(IMAGE_NAME="${WARDEN_IMAGE_REPOSITORY}/${IMAGE_NAME}")
+FULL_PHP_VERSION=""
+MAJOR_VERSION="$(echo "${BUILD_VERSION}" | sed -E 's/([0-9])([0-9])/\1.\2/')"
+# Configure build args specific to this image build
+PHP_VERSION="${MAJOR_VERSION}"
+BUILD_ARGS=(PHP_VERSION)
+BUILD_ARGS+=(IMAGE_NAME="${WARDEN_IMAGE_REPOSITORY}/${IMAGE_NAME}")
 
 docker buildx use warden-builder >/dev/null 2>&1 || docker buildx create --name warden-builder --use
 
@@ -65,18 +66,27 @@ echo "::group::Generating tags for ${IMAGE_NAME}:${BUILD_VERSION} (${BUILD_VARIA
   TAG_SUFFIX="$(echo "${BUILD_VARIANT}" | sed -E 's/^(cli$|cli-)//')"
   [[ ${TAG_SUFFIX} ]] && TAG_SUFFIX="-${TAG_SUFFIX}"
 
-  # Fetch the precise php version from the built image and tag it
-  MINOR_VERSION="$(docker run --rm -t --entrypoint php "${IMAGE_NAME}:build" -r 'echo phpversion();')"
+  if [[ -z "${FULL_PHP_VERSION:-}" ]]; then
+    # Fetch the precise php version from the built image and tag it
+    FULL_PHP_VERSION="$(docker run --rm -t --entrypoint php "${IMAGE_NAME}:build" -r 'echo phpversion();')"
+
+    echo "full_php_version=${FULL_PHP_VERSION}" >> $GITHUB_OUTPUT
+
+    mkdir -p "${PHP_VERSIONS_DIR}"
+    jq -n --arg major "$MAJOR_VERSION" --arg full "$FULL_PHP_VERSION" '{ ($major): $full }' > "${PHP_VERSIONS_DIR}/${MAJOR_VERSION}.json"
+  else
+    echo "Full PHP Version: ${FULL_PHP_VERSION}"
+  fi
 
   # Generate array of tags for the image being built
   IMAGE_TAGS=(
     "${MAJOR_VERSION}${TAG_SUFFIX}"
-    "${MINOR_VERSION}${TAG_SUFFIX}"
+    "${FULL_PHP_VERSION}${TAG_SUFFIX}"
   )
 
 echo "::endgroup::"
 
-echo "::group::Pushing layers to registries for ${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX} (${PLATFORM})"
+echo "::group::Pushing layers to registries for ${IMAGE_NAME}:${FULL_PHP_VERSION}${TAG_SUFFIX} (${PLATFORM})"
 
   # Iterate and push image tags to remote registry
   if [[ ${PUSH_FLAG} != 0 ]]; then
@@ -106,7 +116,7 @@ echo "::group::Pushing layers to registries for ${IMAGE_NAME}:${MINOR_VERSION}${
     mkdir -p "${METADATA_DIR}"
     echo "${JSON}" > "${METADATA_DIR}/${BUILD_VERSION}-${BUILD_VARIANT}-${PLATFORM//\//-}.json"
 
-    echo "::notice title=Container image digest for ${IMAGE_NAME}:${MINOR_VERSION}${TAG_SUFFIX} (${PLATFORM##*/})::${digest}"
+    echo "::notice title=Container image digest for ${IMAGE_NAME}:${FULL_PHP_VERSION}${TAG_SUFFIX} (${PLATFORM##*/})::${digest}"
   fi
 
 echo "::endgroup::"
